map component with eslint stylerules
