--- frontend/src/components/map.html
@@ -1,96 +1,0 @@
-<!-- AIzaSyBdDmKkxJcu6FlyK0Z9SOEEGEMk9U_Daig 
-unused, mostly notes for the react component
-GPS coordinats for 10 markers from geojson-->
-<!DOCTYPE html>
-<html>
-
-<head>
-    <meta charset="UTF-8">
-    <script async defer src="https://maps.googleapis.com/maps/api/js?key=AIzaSyBdDmKkxJcu6FlyK0Z9SOEEGEMk9U_Daig">
-    </script>
-    <style>
-        #map {
-            height: 90vh;
-            width: 100%;
-        }
-
-        #addMark {
-            display: inline-block;
-            width: 100px;
-            height: 40px;
-            background-color: orange;
-        }
-
-        h3 {
-            display: inline-block;
-        }
-    </style>
-</head>
-
-<body>
-    <h3>My Google Maps Demo</h3>
-    <button id="addMark">addMarker</button>
-    <div id="map"></div>
-    <script>
-        var uluru = { lat: 63.387075002372903, lng: 10.3277250005425 }
-
-        var map = new google.maps.Map(document.getElementById('map'), {
-            zoom: 16,
-            center: uluru
-        });
-        //geolocation for the initial place
-        //    var uluru = { lat: 63.387075002372903, lng: 10.3277250005425 }
-        //    var fakeUluru = { lat: 63.387441999029399, lng: 10.3290930003037 };;
-
-        var startLats = [63.387075002372903, 63.387691997704202, 63.387642998353499, 63.387514999582102, 63.387520001503702, 63.398216997487999, 63.398066997137299, 63.398084999471202];
-        var startLongs = [10.3277250005425, 10.3290819995141, 10.3282330021124, 10.328351999716901, 10.3280969991206, 10.3594619979811, 10.3578219979658, 10.357628000185899];
-        var endLats = [63.387441999029399, 63.387777001722696, 63.387514999582102, 63.387471999099603, 63.387520001503702, 63.398066997137299, 63.398084999471202, 63.398135001497998];
-        var endLongs = [10.3290930003037, 10.328826998917799, 10.328351999716901, 10.3280570009369, 10.3283469977953, 10.3578219979658, 10.357628000185899, 10.357228001160401];
-        var amountOfMarkers = 0;
-
-        function addMarker() {
-            console.log("click")
-            var uluru = { lat: startLats[amountOfMarkers], lng: startLongs[amountOfMarkers] };
-            console.log(uluru);
-            var marker = new google.maps.Marker({
-                position: uluru,
-                map: this.map
-            });
-            amountOfMarkers += 1;
-        }
-
-        var addButton = document.getElementById("addMark");
-        addButton.onclick(addMarker());
-
-    </script>
-</body>
-
-</html>
-
-
-<!-- 
-    https://stackoverflow.com/questions/3059044/google-maps-js-api-v3-simple-multiple-marker-example
-    
-    Zoom Scale
-    20 : 1128.497220
-19 : 2256.994440
-18 : 4513.988880
-17 : 9027.977761
-16 : 18055.955520
-15 : 36111.911040
-14 : 72223.822090
-13 : 144447.644200
-12 : 288895.288400
-11 : 577790.576700
-10 : 1155581.153000
-9  : 2311162.307000
-8  : 4622324.614000
-7  : 9244649.227000
-6  : 18489298.450000
-5  : 36978596.910000
-4  : 73957193.820000
-3  : 147914387.600000
-2  : 295828775.300000
-1  : 591657550.500000
-
--->
