--- frontend/src/index.js
@@ -1,8 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./index.css";
-import App from "./App";
-import registerServiceWorker from "./registerServiceWorker";
-
-ReactDOM.render(<App />, document.getElementById("root"));
-registerServiceWorker();
